Move positional modifications to HitCirclePlacementMask
